utils/Rasa/instruction_training_platform/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.24.0
sqlalchemy==1.4.0
pydantic==1.10.9
python-multipart==0.0.6
requests==2.31.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
psycopg2-binary==2.9.9
rasa==3.6.21
tensorflow==2.12.0
jieba==0.42.1
pandas==2.1.4
numpy==1.23.4
scikit-learn==1.1.3
chardet==5.0.0
openpyxl==3.1.2
xlrd==2.0.1
psutil==5.9.0

